DQN

Documentation for DQN.

DQN.agent

DQN.set_greediness!

DQN.set_q_function

DQN.train_step!

DQN.ϵ_greedy

DQN.agent — Method

agent(environment::DiscreteEnvironment, agentParams::AgentParameter)

Train the agent in the given environment with the given parameters.

DQN.set_greediness! — Method

set_greediness!(aP::AgentParameter, step_counter::Int64)

Set the greediness of the agent.

DQN.set_q_function — Method

set_q_function(state_size, action_size)

Create a Q-function with the given state and action size.

DQN.train_step! — Method

train_step!(S, A, R, S´, T, Qϕ, Qϕ´, opt_critic, ap::Parameter, ep::EnvParameter)

Train the critic network with gradient descent.

DQN.ϵ_greedy — Method

ϵ_greedy(Qϕ, s, aP::AgentParameter, eP::EnvParameter)

Return the action to take based on the Q-function and the current state.