DQN

Documentation for DQN.

This implementation of DQN is based on the paper Playing Atari with Deep Reinforcement Learning by Mnih et al. (2013). It solves the following environments:

Acrobot

Lunar Lander

Installation

In the julia REPL, run

using Pkg
Pkg.add(url="https://github.com/SvenDuve/DQN.jl")

This package is not for general use. To be fully functional, it requires the following packages:

RLTypes

Conda

PyCall

Gymnasium, this is a slightly adapted version. The original package can be found here

Run the following:

using Pkg
Pkg.add(url="https://github.com/SvenDuve/RLTypes.jl")
Pkg.add("Conda")
Pkg.add("PyCall")

Then again within julia,

using Conda
using PyCall
Conda.pip("install", "git+https://github.com/SvenDuve/Gymnasium")

Bring package into scope with

using DQN

Example usage

Once the above is installed, the following code can be used to train an agent:

using DDPG
agent(Acrobot(), AgentParameter(training_episodes=100, batch_size=128))

This function returns a data structure containing a trained agent.

Function Reference

DQN.agent

DQN.set_greediness!

DQN.set_q_function

DQN.train_step!

DQN.ϵ_greedy

DQN.agent — Method

agent(environment::DiscreteEnvironment, agentParams::AgentParameter)

Train the agent in the given environment with the given parameters.

DQN.set_greediness! — Method

set_greediness!(aP::AgentParameter, step_counter::Int64)

Set the greediness of the agent.

DQN.set_q_function — Method

set_q_function(state_size, action_size)

Create a Q-function with the given state and action size.

DQN.train_step! — Method

train_step!(S, A, R, S´, T, Qϕ, Qϕ´, opt_critic, ap::Parameter, ep::EnvParameter)

Train the critic network with gradient descent.

DQN.ϵ_greedy — Method

ϵ_greedy(Qϕ, s, aP::AgentParameter, eP::EnvParameter)

Return the action to take based on the Q-function and the current state.